+++ src/utils/constants/constant.js
@@ -1,7 +1,21 @@
 export const CLEAR_FEEDBACKS = "CLEAR_FEEDBACKS";
 export const TOGGLE_NAVBAR_BUTTON = "TOGGLE_NAVBAR_BUTTON";
 export const LOGOUT = "LOGOUT";
+// The name of key in localStorage
 export const REACT_FE_ACCESS_TOKEN = "OBF_ACCESS_TOKEN";
+/**
+ * PATHS OF APPLICATION
+ */
 export const PATH_HOME = "/";
 export const PATH_DASHBOARD = "/dashboard";
 export const PATH_BANK_ACCOUNT_TRANSACTIONS = "/bank-account-transactions";
+/**
+ * 
+ * REMOTE PAGINATION
+ * 
+ */
+// Are pageNumber of server starting from zero or 1?
+// If true, pageNumber will be set to pageNumber - 1
+export const STARTING_ZERO = true;
+// Key of total elements in remote collection
+export const TOTAL_ELEMENTS = "totalElements";