+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Add `bank-account-transactions` api, reducer, page
 + Check for `content` in items
 + Add Bank Account Transaction Page
++ Add Remote Pagination
 
 ## 0.2.0
 + Add CircleCI